requirements.txt
Django==4.1.13
djangorestframework==3.15.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pandas==2.2.1
djongo==1.3.6
pymongo==3.12.3
pymongo[srv]
awscli==1.33.7
boto3==1.34.125
django-storages==1.14.3
django-rest-passwordreset==1.4.1
djangorestframework-simplejwt==5.3.1